stop the bean replacement policy of JpaTest
By default JpaTest will override any DB settings and run with an embedded H2 database, which is the opposite of what we want here. By telling Spring Boot not override the DB settings we ensure that it uses the properties that we are passing on, ie, the connection to a real PostgreSQL database.